test: 테스트 로직 및 메소드 명 변경
