+++ src/test/java/laddergame/domain/LadderTest.java
@@ -1,0 +1,26 @@
+package laddergame.domain;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.util.List;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+public class LadderTest {
+
+    @DisplayName("사다리높이만큼 라인을 생성한다.")
+    @ParameterizedTest
+    @ValueSource(ints = {1, 2, 3, 4})
+    void create(int value) {
+        // given
+        final LadderHeight height = new LadderHeight(value);
+        final LineSize lineSize = new LineSize(new Names(List.of("pobi", "zeze", "crong", "jk")));
+
+        // when
+        Ladder ladder = Ladder.create(lineSize, height, () -> Point.EXIST);
+
+        // then
+        assertThat(ladder.getLinesState()).hasSize(value);
+    }
+}
