+++ src/test/java/laddergame/domain/NameTest.java
@@ -1,6 +1,5 @@
-package laddergame;
+package laddergame.domain;
 
-import laddergame.domain.Name;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,15 +8,6 @@ import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 public class NameTest {
-
-    @DisplayName("이름의 길이가 5를 넘으면 예외를 발생시킨다.")
-    @Test
-    void invalidNameLength() {
-        Assertions.assertThatThrownBy(() -> new Name("abcdef"))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 이름길이는 5글자를 넘을 수 없습니다.");
-
-    }
 
     @DisplayName("이름을 입력받아 객체를 생성한다.")
     @ParameterizedTest
@@ -30,7 +20,16 @@ public class NameTest {
         Assertions.assertThat(name.getName()).isEqualTo(value);
     }
 
-    @DisplayName("이름값에 빈문자열이나 공백을 입력하면 예외를 발생시킨다.")
+    @DisplayName("이름의 길이가 5를 넘으면 예외를 발생시킨다.")
+    @Test
+    void validateNameLength() {
+        Assertions.assertThatThrownBy(() -> new Name("abcdef"))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("[ERROR] 이름길이는 5글자를 넘을 수 없습니다.");
+
+    }
+
+    @DisplayName("이름에 빈문자열이나 공백을 입력하면 예외를 발생시킨다.")
     @ParameterizedTest
     @ValueSource(strings = {" ", "", "   "})
     @NullSource
