+++ src/test/java/laddergame/service/LadderGameTest.java
@@ -1,12 +1,12 @@
-package laddergame;
+package laddergame.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.ArrayList;
 import java.util.List;
-import laddergame.domain.Point;
-import laddergame.service.LadderGame;
 import laddergame.domain.LadderHeight;
 import laddergame.domain.Names;
+import laddergame.domain.Point;
 import laddergame.domain.Result;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,7 +16,7 @@ public class LadderGameTest {
 
     @DisplayName("사다리 게임이 초기화 된다.")
     @Test
-    void init() {
+    void create() {
         // given
         final List<String> input = List.of("pobi", "honux", "crong", "jk");
 
@@ -31,6 +31,14 @@ public class LadderGameTest {
         // then
         assertThat(result.names()).isEqualTo(input);
         assertThat(result.ladder()).hasSize(5)
-                .isEqualTo(LadderTest.createLadder(List.of(true, false, true), 5));
+                .isEqualTo(createLadder(List.of(true, false, true), 5));
+    }
+
+    private List<List<Boolean>> createLadder(final List<Boolean> line, final int size) {
+        List<List<Boolean>> ladder = new ArrayList<>();
+        for (int i = 0; i < size; i++) {
+            ladder.add(line);
+        }
+        return ladder;
     }
 }
